renamed RelayHooksType in RelayHooksTypes
